requirements.txt
base58==2.1.1
cffi==1.17.1
cryptography==44.0.0
ecdsa==0.19.0
pycparser==2.22
pycryptodome==3.21.0
six==1.17.0
